# backend/schemata/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Test Observer",
    "description": "Test Observer API (see https://github.com/canonical/test_observer)",
    "license": {
      "name": "GNU Affero General Public License v3",
      "url": "https://raw.githubusercontent.com/canonical/test_observer/refs/heads/main/backend/LICENSE"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/artefacts": {
      "get": {
        "tags": [
          "artefacts"
        ],
        "summary": "Get Artefacts",
        "description": "Get latest artefacts optionally by family",
        "operationId": "get_artefacts_v1_artefacts_get",
        "parameters": [
          {
            "name": "family",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/FamilyName"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Family"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArtefactResponse"
                  },
                  "title": "Response Get Artefacts V1 Artefacts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/artefacts/{artefact_id}": {
      "get": {
        "tags": [
          "artefacts"
        ],
        "summary": "Get Artefact",
        "operationId": "get_artefact_v1_artefacts__artefact_id__get",
        "parameters": [
          {
            "name": "artefact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Artefact Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtefactResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "artefacts"
        ],
        "summary": "Patch Artefact",
        "operationId": "patch_artefact_v1_artefacts__artefact_id__patch",
        "parameters": [
          {
            "name": "artefact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Artefact Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArtefactPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtefactResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/artefacts/{artefact_id}/versions": {
      "get": {
        "tags": [
          "artefacts"
        ],
        "summary": "Get Artefact Versions",
        "operationId": "get_artefact_versions_v1_artefacts__artefact_id__versions_get",
        "parameters": [
          {
            "name": "artefact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Artefact Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArtefactVersionResponse"
                  },
                  "title": "Response Get Artefact Versions V1 Artefacts  Artefact Id  Versions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ArtefactPatch": {
        "properties": {
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ArtefactStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "archived": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived"
          },
          "stage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StageName"
              },
              {
                "type": "null"
              }
            ]
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "type": "object",
        "title": "ArtefactPatch"
      },
      "ArtefactResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "track": {
            "type": "string",
            "title": "Track"
          },
          "store": {
            "type": "string",
            "title": "Store"
          },
          "branch": {
            "type": "string",
            "title": "Branch"
          },
          "series": {
            "type": "string",
            "title": "Series"
          },
          "repo": {
            "type": "string",
            "title": "Repo"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "os": {
            "type": "string",
            "title": "Os"
          },
          "release": {
            "type": "string",
            "title": "Release"
          },
          "owner": {
            "type": "string",
            "title": "Owner"
          },
          "sha256": {
            "type": "string",
            "title": "Sha256"
          },
          "image_url": {
            "type": "string",
            "title": "Image Url"
          },
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "family": {
            "type": "string",
            "title": "Family"
          },
          "status": {
            "$ref": "#/components/schemas/ArtefactStatus"
          },
          "comment": {
            "type": "string",
            "title": "Comment"
          },
          "archived": {
            "type": "boolean",
            "title": "Archived"
          },
          "assignee": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "due_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "bug_link": {
            "type": "string",
            "title": "Bug Link"
          },
          "all_environment_reviews_count": {
            "type": "integer",
            "title": "All Environment Reviews Count"
          },
          "completed_environment_reviews_count": {
            "type": "integer",
            "title": "Completed Environment Reviews Count"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "version",
          "track",
          "store",
          "branch",
          "series",
          "repo",
          "source",
          "os",
          "release",
          "owner",
          "sha256",
          "image_url",
          "stage",
          "family",
          "status",
          "comment",
          "archived",
          "assignee",
          "due_date",
          "created_at",
          "bug_link",
          "all_environment_reviews_count",
          "completed_environment_reviews_count"
        ],
        "title": "ArtefactResponse"
      },
      "ArtefactStatus": {
        "type": "string",
        "enum": [
          "APPROVED",
          "MARKED_AS_FAILED",
          "UNDECIDED"
        ],
        "title": "ArtefactStatus"
      },
      "ArtefactVersionResponse": {
        "properties": {
          "version": {
            "type": "string",
            "title": "Version"
          },
          "artefact_id": {
            "type": "integer",
            "title": "Artefact Id"
          }
        },
        "type": "object",
        "required": [
          "version",
          "artefact_id"
        ],
        "title": "ArtefactVersionResponse"
      },
...
      "FamilyName": {
        "type": "string",
        "enum": [
          "snap",
          "deb",
          "charm",
          "image"
        ],
        "title": "FamilyName"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "StageName": {
        "type": "string",
        "enum": [
          "proposed",
          "updates",
          "edge",
          "beta",
          "candidate",
          "stable",
          "pending",
          "current"
        ],
        "title": "StageName"
      },
...
      "UserResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "launchpad_handle": {
            "type": "string",
            "title": "Launchpad Handle"
          },
          "launchpad_email": {
            "type": "string",
            "title": "Launchpad Email"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "launchpad_handle",
          "launchpad_email",
          "name"
        ],
        "title": "UserResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}